Move todo stuff
